style: change import for app
